Update top level structure.
