+++ mbed_cloud/connect.py
@@ -11,16 +11,17 @@
 #   permitted to do so under the terms of a subsisting license agreement
 #   from ARM Limited or its affiliates.
 # --------------------------------------------------------------------------
-"""Public API for Device API."""
+"""Public API for mDS and Statistics APIs."""
 from __future__ import absolute_import
 import base64
 from collections import defaultdict
+import datetime
 import logging
+import re
 from six import iteritems
 from six.moves import queue
 import threading
 import time
-import urllib
 
 # Import common functions and exceptions from frontend API
 from mbed_cloud import BaseAPI
@@ -29,30 +30,24 @@ from mbed_cloud.exceptions import CloudAsyncError
 from mbed_cloud.exceptions import CloudTimeoutError
 from mbed_cloud.exceptions import CloudUnhandledError
 from mbed_cloud.exceptions import CloudValueError
-from mbed_cloud import PaginatedResponse
 
 # Import backend API
-import mbed_cloud._backends.device_catalog as dc
-from mbed_cloud._backends.device_catalog.models import DeviceData
-from mbed_cloud._backends.device_catalog.rest import \
-    ApiException as DeviceCatalogApiException
-import mbed_cloud._backends.device_query_service as dc_queries
-from mbed_cloud._backends.device_query_service.rest import \
-    ApiException as DeviceQueryServiceApiException
 import mbed_cloud._backends.mds as mds
 from mbed_cloud._backends.mds.rest import ApiException as MdsApiException
+import mbed_cloud._backends.statistics as statistics
+from mbed_cloud._backends.statistics.models import Metric as MetricData
+from mbed_cloud._backends.statistics.rest import ApiException as StatisticsApiException
 
 LOG = logging.getLogger(__name__)
 
 
-class DeviceAPI(BaseAPI):
-    """API reference for the Device API.
+class ConnectAPI(BaseAPI):
+    """API reference for the Connect API.
 
     Exposing functionality for doing a range of device related actions:
-        - Listing registered and connected devices
+        - Listing connected devices
         - Exploring and managing resources and resource values on said devices
         - Setup resource subscriptions and webhooks for resource monitoring
-        - Create and manage device queries
     """
 
     def __init__(self, params={}, b64decode=True):
@@ -61,12 +56,10 @@ class DeviceAPI(BaseAPI):
         In addition we need to setup some special handling of background
         threads for the mDS functionality - as it relies on background polling.
         """
-        super(DeviceAPI, self).__init__(params)
+        super(ConnectAPI, self).__init__(params)
 
         # Initialize the wrapped APIs
         self.mds = self._init_api(mds)
-        self.dc = self._init_api(dc)
-        self.dc_queries = self._init_api(dc_queries)
 
         self._db = {}
         self._queues = defaultdict(lambda: defaultdict(queue.Queue))
@@ -77,6 +70,14 @@ class DeviceAPI(BaseAPI):
                                                        mds=self.mds)
         self._long_polling_is_active = False
         self._long_polling_thread.daemon = True
+
+        self.statistics = self._init_api(statistics)
+        # This API is a bit weird, so create the "authorization" string
+        authorization = self.statistics.configuration.api_key['Authorization']
+        self._auth = "Bearer %s" % (authorization,)
+        # API requires to send what fields should be returned in response
+        self._include_all = "registered_devices,transactions,apikeys,"\
+            "bootstraps_successful,bootstraps_failed,bootstraps_pending"
 
     def start_notifications(self):
         """Start the long-polling thread.
@@ -439,273 +440,54 @@ class DeviceAPI(BaseAPI):
         self._queues.clear()
         return
 
-    @catch_exceptions(DeviceCatalogApiException)
-    def list_devices(self, **kwargs):
-        """List devices in the device catalog.
+    @catch_exceptions(StatisticsApiException)
+    def get_metrics(self, include=None, interval="1d", **kwargs):
+        """Get statistics.
 
-        Example usage, listing all registered devices in the catalog:
-
-        .. code-block:: python
-
-            filters = { 'state': 'registered' }
-            devices = api.list_devices(order='asc', filters=filters)
-            for d, idx in devices.iteritems():
-                print(idx, d.id)
-
-        :param int limit: (Optional) The number of devices to retrieve.
-        :param str order: (Optional) The ordering direction, ascending (asc) or
-            descending (desc)
-        :param str after: (Optional) Get devices after/starting at given `device_id`
-        :param filters: (Optional) Dictionary of filters to apply.
-        :returns: a list of :py:class:`Device` objects registered in the catalog.
-        :rtype: PaginatedResponse
+        :param str include: What fields to include in response. None will return all.
+        :param str interval: Group data by this interval in days, weeks or hours.
+            Sample values: 2h, 3w, 4d.
+        :param datetime start: Fetch the data with timestamp greater than or equal to this value.
+            The parameter is not mandatory, if the period is specified.
+        :param datetime end: Fetch the data with timestamp less than this value.
+            The parameter is not mandatory, if the period is specified.
+        :param str period: Period. Fetch the data for the period in days, weeks or hours.
+            Sample values: 2h, 3w, 4d.
+            The parameter is not mandatory, if the start and end time are specified
+        :returns: a list of :py:class:`Metric` objects.
         """
-        kwargs = self._verify_sort_options(kwargs)
-        kwargs = self._verify_filters(kwargs)
+        if not include:
+            include = self._include_all
+        self._verify_arguments(interval, kwargs)
+        api = self.statistics.StatisticsApi()
+        return api.v3_metrics_get(include, interval, self._auth, **kwargs).data
 
-        api = self.dc.DefaultApi()
-        return PaginatedResponse(api.device_list, lwrap_type=Device, **kwargs)
+    @catch_exceptions(StatisticsApiException)
+    def get_account_metrics(self, include=None, interval="1d", **kwargs):
+        """Get account-specific statistics.
 
-    @catch_exceptions(DeviceCatalogApiException)
-    def get_device(self, device_id):
-        """Get device details from catalog.
-
-        :param device_id: the ID of the device to retrieve (str)
-        :returns: device object matching the `device_id`.
-        :rtype: Device
+        :param str include: What fields to include in response. None will return all.
+        :param str interval: Group data by this interval in days, weeks or hours.
+            Sample values: 2h, 3w, 4d.
+        :param datetime start: Fetch the data with timestamp greater than or equal to this value.
+            The parameter is not mandatory, if the period is specified.
+        :param datetime end: Fetch the data with timestamp less than this value.
+            The parameter is not mandatory, if the period is specified.
+        :param str period: Period. Fetch the data for the period in days, weeks or hours.
+            Sample values: 2h, 3w, 4d.
+            The parameter is not mandatory, if the start and end time are specified
+        :returns: a list of :py:class:`Metric` objects.
         """
-        api = self.dc.DefaultApi()
-        return Device(api.device_retrieve(device_id))
-
-    @catch_exceptions(DeviceCatalogApiException)
-    def update_device(self, device_id, **kwargs):
-        """Update existing device in catalog.
-
-        .. code-block:: python
-
-            existing_device = api.get_device(...)
-            updated_device = api.update_device(
-                existing_device.id,
-                certificate_fingerprint = "something new"
-            )
-
-        :param bool auto_update: Mark this device for auto firmware update
-        :param obj custom_attributes: Up to 5 custom JSON attributes
-        :param str description: The description of the device
-        :param str name: The name of the device
-        :param str alias: The alias of the device
-        :param str certificate_fingerprint: Fingerprint of the device certificate
-        :param str certificate_issuer_id: ID of the issuer of the certificate
-        :returns: the updated device object
-        :rtype: Device
-        """
-        kwargs = Device._verify_args(kwargs)
-        api = self.dc.DefaultApi()
-        body = self.dc.DeviceDataPostRequest(**kwargs)
-        return Device(api.device_update(device_id, body))
-
-    @catch_exceptions(DeviceCatalogApiException)
-    def add_device(self, **kwargs):
-        """Add a new device to catalog.
-
-        .. code-block:: python
-
-            device = {
-                "mechanism": "connector",
-                "certificateFingerprint": "<certificate>",
-                "name": "New device name",
-                "auto_update": True,
-                "certificateIssuerId": "<id>"
-            }
-            resp = api.add_device(**device)
-            print(resp.created_at)
-
-        :param str account_id: The owning IAM account ID
-        :param bool auto_update: Mark this device for auto firmware update
-        :param obj custom_attributes: Up to 5 custom JSON attributes
-        :param str deployed_state: State of the device's deployment
-        :param str description: The description of the device
-        :param str device_class: Class of the device
-        :param str id: The ID of the device
-        :param str manifest_url: URL for the current device manifest
-        :param str mechanism: The ID of the channel used to communicate with the device
-        :param str mechanism_url: The address of the connector to use
-        :param str name: The name of the device
-        :param str serial_number: The serial number of the device
-        :param str state: The current state of the device
-        :param int trust_class: The device trust class
-        :param int trust_level: The device trust level
-        :param str vendor_id: The device vendor ID
-        :param str alias: The alias of the device
-        :param datetime bootstrap_certificate_expiration:
-        :param str certificate_fingerprint: Fingerprint of the device certificate
-        :param str certificate_issuer_id: ID of the issuer of the certificate
-        :param datetime connector_certificate_expiration: Expiration date of the certificate
-        used to connect to connector server
-        :param int device_execution_mode: The device class
-        :param str firmware_checksum: The SHA256 checksum of the current firmware image
-        :param datetime manifest_timestamp: The timestamp of the current manifest version
-        :return: the newly created device object.
-        :rtype: Device
-        """
-        kwargs = Device._verify_args(kwargs)
-        api = self.dc.DefaultApi()
-        device = DeviceData(**kwargs)
-        return Device(api.device_create(device))
-
-    @catch_exceptions(DeviceCatalogApiException)
-    def delete_device(self, id):
-        """Delete device from catalog.
-
-        :param str id: ID of device in catalog to delete
-        :return: void
-        """
-        api = self.dc.DefaultApi()
-        return api.device_destroy(id=id)
-
-    @catch_exceptions(DeviceQueryServiceApiException)
-    def list_queries(self, **kwargs):
-        """List queries in device query service.
-
-        :param int limit: (Optional) The number of devices to retrieve.
-        :param str order: (Optional) The ordering direction, ascending (asc) or
-            descending (desc)
-        :param str after: (Optional) Get devices after/starting at given `device_id`
-        :returns: a list of :py:class:`Query` objects.
-        :rtype: PaginatedResponse
-        """
-        kwargs = self._verify_sort_options(kwargs)
-        api = self.dc_queries.DefaultApi()
-
-        return PaginatedResponse(api.device_query_list, lwrap_type=Query, **kwargs)
-
-    @catch_exceptions(DeviceQueryServiceApiException)
-    def add_query(self, name, filter, custom_attributes=None, **kwargs):
-        """Add a new query to device query service.
-
-        .. code-block:: python
-
-            f = api.add_query(
-                name = "Query name",
-                query = {},
-                custom_attributes = {
-                    "foo": "bar"
-                }
-            )
-            print(f.created_at)
-
-        :param str name: Name of query
-        :param dict filter: Filter properties to apply
-        :param dict custom_attributes: Extra query attributes
-        :param return: the newly created query object.
-        :return: the newly created query object
-        :rtype: Query
-        """
-        api = self.dc_queries.DefaultApi()
-
-        # Ensure we have the correct types and get the new query object based on
-        # passed in query object and custom attributes.
-        query = self._get_query_attributes(filter, custom_attributes)
-
-        # Create the query object
-        f = self.dc_queries.DeviceQuery(name=name, query=query, **kwargs)
-
-        return Query(api.device_query_create(f))
-
-    @catch_exceptions(DeviceQueryServiceApiException)
-    def update_query(self, query_id, name, filter, custom_attributes=None, **kwargs):
-        """Update existing query in device query service.
-
-        .. code-block:: python
-
-            q = api.get_query(...)
-            new_custom_attributes = {
-                "foo": "bar"
-            }
-            new_q = api.update_query(
-                query_id = q.id,
-                name = "new name",
-                filter = q.filter,
-                custom_attributes = new_custom_attributes
-            )
-
-        :param str query_id: Existing query ID to update
-        :param str name: (New) name of query
-        :param dict query: (New) query properties to apply
-        :param dict custom_attributes: (New) extra query attributes
-        :param return: the newly updated query object.
-        :rtype: Query
-        """
-        api = self.dc_queries.DefaultApi()
-
-        # Get urlencoded query attribute
-        query = self._get_query_attributes(filter, custom_attributes)
-
-        body = self.dc_queries.DeviceQueryPostPutRequest(
-            name=name,
-            query=query,
-            **kwargs
-        )
-
-        return Query(api.device_query_update(query_id, body))
-
-    @catch_exceptions(DeviceQueryServiceApiException)
-    def delete_query(self, query_id):
-        """Delete query in device query service.
-
-        :param int query_id: id of the query to delete
-        :param return: void
-        """
-        api = self.dc_queries.DefaultApi()
-        api.device_query_destroy(query_id)
-        return
-
-    @catch_exceptions(DeviceQueryServiceApiException)
-    def get_query(self, query_id):
-        """Get query in device query service.
-
-        :param int query_id: id of the query to get
-        :returns: device query object
-        :rtype: Query
-        """
-        api = self.dc_queries.DefaultApi()
-        return Query(api.device_query_retrieve(query_id))
+        if not include:
+            include = self._include_all
+        self._verify_arguments(interval, kwargs)
+        api = self.statistics.AccountApi()
+        return api.v3_metrics_get(include, interval, self._auth, **kwargs).data
 
     def _subscription_handler(self, queue, callback_fn):
         while True:
             value = queue.get()
             callback_fn(value)
-
-    def _get_query_attributes(self, query, custom_attributes):
-        # Ensure the query is of dict type
-        if query and not isinstance(query, dict):
-            raise CloudValueError("'query' parameter needs to be of type dict")
-
-        # Add custom attributes, if provided
-        if custom_attributes:
-            if not isinstance(custom_attributes, dict):
-                raise CloudValueError("Custom attributes when creating query"
-                                      "needs to be dict object")
-            for k, v in custom_attributes.iteritems():
-                if not k:
-                    LOG.warning("Ignoring custom attribute with value %r as key is empty" % (v,))
-                    continue
-                query['custom_attributes__' + k] = v
-
-        # Ensure query is valid
-        if not query.keys():
-            raise CloudValueError("'query' parameter not valid, needs to contain query keys")
-
-        return self._urlify_query(query)
-
-    def _urlify_query(self, query):
-        # Quote strings using %20, not '+' which is default when urlencoding dicts
-        for k, v in query.iteritems():
-            if type(v) is str:
-                query[k] = urllib.quote(v)
-
-        # Encode the query string
-        return urllib.urlencode(query)
 
     def _get_value_synchronized(self, consumer, timeout=None):
         start_time = int(time.time())
@@ -726,6 +508,33 @@ class DeviceAPI(BaseAPI):
             raise CloudAsyncError(consumer.error)
 
         return consumer.value
+
+    def _convert_to_UTC_RFC3339(self, time, name):
+        if not isinstance(time, datetime.datetime):
+            raise CloudValueError("%s should be of type datetime" % (name))
+        return time.isoformat() + "Z"
+
+    def _verify_arguments(self, interval, kwargs):
+        start = kwargs.get("start", None)
+        end = kwargs.get("end", None)
+        period = kwargs.get("period", None)
+        if not start and not end and not period:
+            raise CloudValueError("start and end is mandatory if period is not specified.")
+        if start and not end:
+            raise CloudValueError("end is required if start is specified.")
+        if end and not start:
+            raise CloudValueError("start is required if end is specified.")
+        pattern = re.compile("[0-9]+[h|d|w]$")
+        if period and not pattern.match(period):
+            raise CloudValueError("period is incorrect. Sample values: 2h, 3w, 4d.")
+        if interval and not pattern.match(interval):
+            raise CloudValueError("interval is incorrect. Sample values: 2h, 3w, 4d.")
+        # convert start into UTC RFC3339 format
+        if start:
+            kwargs['start'] = self._convert_to_UTC_RFC3339(start, 'start')
+        # convert end into UTC RFC3339 format
+        if end:
+            kwargs['end'] = self._convert_to_UTC_RFC3339(end, 'end')
 
 
 class AsyncConsumer(object):
@@ -939,435 +748,6 @@ class ConnectedDevice(object):
         return str(self.to_dict())
 
 
-class Device(object):
-    """Describes device object from the catalog."""
-
-    def __init__(self, device_obj):
-        """Override __init__ and allow passing in backend object."""
-        if not isinstance(device_obj, dict):
-            device_obj = device_obj.to_dict()
-        for key, value in iteritems(Device._get_map_attributes()):
-            setattr(self, "_%s" % key, device_obj.get(value, None))
-
-    @staticmethod
-    def _get_map_attributes():
-        return {
-            "account_id": "account_id",
-            "auto_update": "auto_update",
-            "bootstrapped_timestamp": "bootstrapped_timestamp",
-            "created_at": "created_at",
-            "custom_attributes": "custom_attributes",
-            "deployed_state": "deployed_state",
-            "last_deployment": "deployment",
-            "description": "description",
-            "device_class": "device_class",
-            "id": "id",
-            "manifest_url": "manifest",
-            "mechanism": "mechanism",
-            "mechanism_url": "mechanism_url",
-            "name": "name",
-            "serial_number": "serial_number",
-            "state": "state",
-            "trust_class": "trust_class",
-            "trust_level": "trust_level",
-            "updated_at": "updated_at",
-            "vendor_id": "vendor_id",
-            "alias": "endpoint_name",
-            "bootstrap_certificate_expiration": "bootstrap_expiration_date",
-            "certificate_fingerprint": "device_key",
-            "certificate_issuer_id": "ca_id",
-            "connector_certificate_expiration": "connector_expiration_date",
-            "device_execution_mode": "device_execution_mode",
-            "firmware_checksum": "firmware_checksum",
-            "manifest_timestamp": "manifest_timestamp"
-        }
-
-    @staticmethod
-    def _verify_args(kwargs):
-        """Verify arguments before sending to Backend API."""
-        for key, value in iteritems(Device._get_map_attributes()):
-            if key in kwargs:
-                kwargs[value] = kwargs[key]
-                del kwargs[key]
-        return kwargs
-
-    @property
-    def account_id(self):
-        """The owning IAM account ID.
-
-        :rtype: str
-        """
-        return self._account_id
-
-    @property
-    def auto_update(self):
-        """Mark this device for auto firmware update.
-
-        :rtype: bool
-        """
-        return self._auto_update
-
-    @property
-    def bootstrapped_timestamp(self):
-        """The time device was created..
-
-        :rtype: datetime
-        """
-        return self._bootstrapped_timestamp
-
-    @property
-    def created_at(self):
-        """The time the device was created.
-
-        :rtype: datetime
-        """
-        return self._created_at
-
-    @property
-    def custom_attributes(self):
-        """Up to 5 custom JSON attributes.
-
-        :rtype: object
-        """
-        return self._custom_attributes
-
-    @property
-    def deployed_state(self):
-        """State of the device's deployment.
-
-        :rtype: str
-        """
-        return self._deployed_state
-
-    @property
-    def last_deployment(self):
-        """The last deployment used on the device.
-
-        :rtype: datetime
-        """
-        return self._last_deployment
-
-    @property
-    def description(self):
-        """The description of the device.
-
-        :rtype: str
-        """
-        return self._description
-
-    @property
-    def device_class(self):
-        """Class of the device.
-
-        :rtype: str
-        """
-        return self._device_class
-
-    @property
-    def id(self):
-        """The ID of the device.
-
-        :rtype: str
-        """
-        return self._id
-
-    @property
-    def manifest_url(self):
-        """URL for the current device manifest.
-
-        :rtype: str
-        """
-        return self._manifest_url
-
-    @property
-    def mechanism(self):
-        """The ID of the channel used to communicate with the device.
-
-        :rtype: str
-        """
-        return self._mechanism
-
-    @property
-    def mechanism_url(self):
-        """The address of the connector to use.
-
-        :rtype: str
-        """
-        return self._mechanism_url
-
-    @property
-    def name(self):
-        """The name of the device.
-
-        :rtype: str
-        """
-        return self._name
-
-    @property
-    def serial_number(self):
-        """The serial number of the device.
-
-        :rtype: str
-        """
-        return self._serial_number
-
-    @property
-    def state(self):
-        """The current state of the device.
-
-        :rtype: str
-        """
-        return self._state
-
-    @property
-    def trust_class(self):
-        """The device trust class.
-
-        The time the object was created
-
-        :return: The created_at of this Query.
-        :rtype: int
-        """
-        return self._trust_class
-
-    @property
-    def trust_level(self):
-        """The device trust level.
-
-        The time the object was created
-
-        :return: The created_at of this Query.
-        :rtype: int
-        """
-        return self._trust_level
-
-    @property
-    def updated_at(self):
-        """The time the device was updated.
-
-        The time the object was created
-
-        :return: The created_at of this Query.
-        :rtype: datetime
-        """
-        return self._updated_at
-
-    @property
-    def vendor_id(self):
-        """The device vendor ID.
-
-        The time the object was created
-
-        :return: The created_at of this Query.
-        :rtype: str
-        """
-        return self._vendor_id
-
-    @property
-    def alias(self):
-        """The alias of the device.
-
-        :rtype: str
-        """
-        return self._alias
-
-    @property
-    def bootstrap_certificate_expiration(self):
-        """Expiration date of certificate.
-
-        :rtype: datetime
-        """
-        return self._bootstrap_certificate_expiration
-
-    @property
-    def certificate_fingerprint(self):
-        """Fingerprint of the device certificate.
-
-        :rtype: str
-        """
-        return self._certificate_fingerprint
-
-    @property
-    def certificate_issuer_id(self):
-        """ID of the issuer of the certificate.
-
-        :rtype: str
-        """
-        return self._certificate_issuer_id
-
-    @property
-    def connector_certificate_expiration(self):
-        """Expiration date of the certificate used to connect to connector server.
-
-        :rtype: datetime
-        """
-        return self._connector_certificate_expiration
-
-    @property
-    def device_execution_mode(self):
-        """The device class.
-
-        :rtype: int
-        """
-        return self._device_execution_mode
-
-    @property
-    def firmware_checksum(self):
-        """The SHA256 checksum of the current firmware image.
-
-        :rtype: str
-        """
-        return self._firmware_checksum
-
-    @property
-    def manifest_timestamp(self):
-        """The timestamp of the current manifest version.
-
-        :rtype: datetime
-        """
-        return self._manifest_timestamp
-
-    def to_dict(self):
-        """Return dictionary of object."""
-        return {
-            "account_id": self._account_id,
-            "auto_update": self._auto_update,
-            "bootstrapped_timestamp": self._bootstrapped_timestamp,
-            "created_at": self._created_at,
-            "custom_attributes": self._custom_attributes,
-            "deployed_state": self._deployed_state,
-            "last_deployment": self._last_deployment,
-            "description": self._description,
-            "device_class": self._device_class,
-            "id": self._id,
-            "manifest_url": self._manifest_url,
-            "mechanism": self._mechanism,
-            "mechanism_url": self._mechanism_url,
-            "name": self._name,
-            "serial_number": self._serial_number,
-            "state": self._state,
-            "trust_class": self._trust_class,
-            "trust_level": self._trust_level,
-            "updated_at": self._updated_at,
-            "vendor_id": self._vendor_id,
-            "alias": self._alias,
-            "bootstrap_certificate_expiration": self._bootstrap_certificate_expiration,
-            "certificate_fingerprint": self._certificate_fingerprint,
-            "certificate_issuer_id": self._certificate_issuer_id,
-            "connector_certificate_expiration": self._connector_certificate_expiration,
-            "device_execution_mode": self._device_execution_mode,
-            "firmware_checksum": self._firmware_checksum,
-            "manifest_timestamp": self._manifest_timestamp
-        }
-
-    def __repr__(self):
-        """For print and pprint."""
-        return str(self.to_dict())
-
-
-class Query(object):
-    """Describes device query object."""
-
-    def __init__(self, query_obj):
-        """Override __init__ and allow passing in backend object."""
-        if not isinstance(query_obj, dict):
-            query_obj = query_obj.to_dict()
-        for key, value in iteritems(Query._get_map_attributes()):
-            setattr(self, "_%s" % key, query_obj.get(value, None))
-
-    @staticmethod
-    def _get_map_attributes():
-        return {
-            "created_at": "created_at",
-            "description": "description",
-            "id": "id",
-            "name": "name",
-            "updated_at": "updated_at",
-            "filter": "query"
-        }
-
-    @property
-    def created_at(self):
-        """Get the created_at of this Query.
-
-        The time the object was created
-
-        :return: The created_at of this Query.
-        :rtype: datetime
-        """
-        return self._created_at
-
-    @property
-    def description(self):
-        """Get the description of this Query.
-
-        The description of the object
-
-        :return: The description of this Query.
-        :rtype: str
-        """
-        return self._description
-
-    @property
-    def id(self):
-        """Get the id of this Query.
-
-        The ID of the query
-
-        :return: The id of this Query.
-        :rtype: str
-        """
-        return self._id
-
-    @property
-    def name(self):
-        """Get the name of this Query.
-
-        The name of the query
-
-        :return: The name of this Query.
-        :rtype: str
-        """
-        return self._name
-
-    @property
-    def updated_at(self):
-        """Get the updated_at of this Query.
-
-        The time the object was updated
-
-        :return: The updated_at of this Query.
-        :rtype: datetime
-        """
-        return self._updated_at
-
-    @property
-    def filter(self):
-        """Get the query of this Query.
-
-        The device query
-
-        :return: The query of this Query.
-        :rtype: str
-        """
-        return self._filter
-
-    def to_dict(self):
-        """Return dictionary of object."""
-        return {
-            "created_at": self.created_at,
-            "description": self.description,
-            "id": self.id,
-            "name": self.name,
-            "updated_at": self.updated_at,
-            "filter": self.filter
-        }
-
-    def __repr__(self):
-        """For print and pprint."""
-        return str(self.to_dict())
-
-
 class Webhook(object):
     """Describes webhook object."""
 
@@ -1492,3 +872,11 @@ class Resource(object):
     def __repr__(self):
         """For print and pprint."""
         return str(self.to_dict())
+
+
+class Metric(MetricData):
+    """Describes Metric object from statistics."""
+
+    def __init__(self, data_obj):
+        """Override __init__ and allow passing in backend object."""
+        super(Metric, self).__init__(**data_obj.to_dict())
